Added the 'barbero_mode-transitions' experiments, along with its 'fsm_eigenvalue' data files
It's built upon the base 'barbero' experiments and it focuses on the mode
transitions (local-local and local-global), while also substantially increasing
the 'strip length' resolution, by setting its step size to 0.01 mm (10 um,
which is roughly the size of a human red blood cell).

This enables us to model the structures behavior around its mode transitions in
fine detail.